# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.ApplyStatusResponse:
    properties:
      applied_at:
        type: string
      catatan_tolak:
        type: string
      jenis_ujikom:
        type: string
      status:
        description: menunggu_verifikasi | lolos | tidak_lolos
        type: string
      status_kode:
        description: ujikom_menunggu_verifikasi, etc
        type: string
    type: object
  entity.ApplyUjikomRequest:
    properties:
      jenis_ujikom:
        type: string
    required:
    - jenis_ujikom
    type: object
...
  entity.CBTListQuestionsResponse:
    properties:
      attempt_id:
        type: string
      questions:
        items:
          $ref: '#/definitions/entity.CBTQuestionItem'
        type: array
      total:
        type: integer
    type: object
  entity.CBTQuestionItem:
    properties:
      num:
        type: integer
      options:
        items:
          $ref: '#/definitions/entity.CBTQuestionOption'
        type: array
      question_id:
        type: string
      question_text:
        type: string
      type:
        type: string
      weight:
        type: number
    type: object
  entity.CBTQuestionOption:
    properties:
      id:
        type: string
      option_key:
        type: string
      option_text:
        type: string
    type: object
...
  entity.CBTSubmitResponse:
    properties:
      attempt_id:
        type: string
      score:
        type: number
      submitted_at:
        type: string
    type: object
  entity.CreateExamRequest:
    properties:
      code:
        type: string
      durasi_menit:
        type: integer
      jadwal_mulai:
        type: string
      jadwal_selesai:
        type: string
      name:
        type: string
      package_ids:
        items:
          type: string
        type: array
      participant_ids:
        items:
          type: string
        type: array
      question_ids:
        items:
          type: string
        type: array
      shuffle_questions:
        type: boolean
      tampilkan_leaderboard:
        type: boolean
    required:
    - code
    - durasi_menit
    - jadwal_mulai
    - jadwal_selesai
    - name
    type: object
...
  entity.DokumenPersyaratanItem:
    properties:
      batasan:
        type: string
      deskripsi:
        type: string
      id:
        type: string
      kode:
        type: string
      nama:
        type: string
      tipeInput:
        description: file | text_portofolio | text_essay
        type: string
      untukJenisUjikom:
        type: string
      urutan:
        type: integer
    type: object
  entity.ExamContentInfo:
    properties:
      sort_order:
        type: integer
      source_id:
        type: string
      source_type:
        type: string
    type: object
  entity.ExamDetailResponse:
    properties:
      code:
        type: string
      contents:
        items:
          $ref: '#/definitions/entity.ExamContentInfo'
        type: array
      created_at:
        type: string
      durasi_menit:
        type: integer
      id:
        type: string
      jadwal_mulai:
        type: string
      jadwal_selesai:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/entity.ExamParticipantInfo'
        type: array
      shuffle_questions:
        type: boolean
      status:
        type: string
      tampilkan_leaderboard:
        type: boolean
      updated_at:
        type: string
      verification_status:
        type: string
    type: object
  entity.ExamListItem:
    properties:
      code:
        type: string
      created_at:
        type: string
      durasi_menit:
        type: integer
      id:
        type: string
      jadwal_mulai:
        type: string
      jadwal_selesai:
        type: string
      name:
        type: string
      participant_count:
        type: integer
      status:
        type: string
      verification_status:
        type: string
    type: object
  entity.ExamParticipantInfo:
    properties:
      user_id:
        type: string
      user_name:
        type: string
    type: object
...
  entity.GetExamListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.ExamListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  entity.ListDokumenPersyaratanResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.DokumenPersyaratanItem'
        type: array
    type: object
  entity.LoginRequest:
    properties:
      identifier:
        description: email atau username
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  entity.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/entity.UserResponse'
    type: object
...
  entity.RoleInfo:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.UserResponse:
    properties:
      default_home_path:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/entity.RoleInfo'
        type: array
      username:
        type: string
    type: object
  helper.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service (hexagon/clean architecture).
  title: Widyaprada Backend API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan identifier (email/username) dan password
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/v1/cbt/attempts/{attemptId}/questions:
    get:
      consumes:
      - application/json
      description: Daftar soal untuk attempt
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTListQuestionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar soal
      tags:
      - cbt
  /api/v1/cbt/attempts/{attemptId}/questions/{num}:
    get:
      consumes:
      - application/json
      description: Get soal by nomor (1-based)
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      - description: Nomor soal
        in: path
        name: num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTQuestionItem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soal per nomor
      tags:
      - cbt
  /api/v1/cbt/attempts/{attemptId}/submit:
    post:
      consumes:
      - application/json
      description: Submit ujian dan hitung skor (PG/B-S otomatis)
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CBTSubmitResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit ujian
      tags:
      - cbt
...
  /api/v1/exams:
    get:
      consumes:
      - application/json
      description: Get paginated list of exams with filters
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      - description: Draft, Diterbitkan, Berlangsung, Selesai, all
        in: query
        name: status
        type: string
      - description: Belum, Sudah, all
        in: query
        name: status_verifikasi
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetExamListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List exams
      tags:
      - exams
    post:
      consumes:
      - application/json
      description: Create new exam (Draft status)
      parameters:
      - description: Create exam request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateExamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.ExamDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create exam
      tags:
      - exams
...
  /api/v1/ujikom/apply:
    post:
      consumes:
      - application/json
      description: Apply pendaftaran dengan dokumen persyaratan
      parameters:
      - description: Apply request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.ApplyUjikomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Apply pendaftaran Ujikom
      tags:
      - ujikom
  /api/v1/ujikom/apply/status:
    get:
      consumes:
      - application/json
      description: Menunggu verifikasi / lolos / tidak lolos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ApplyStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Status pendaftaran user
      tags:
      - ujikom
  /api/v1/ujikom/dokumen-persyaratan:
    get:
      consumes:
      - application/json
      description: Daftar dokumen persyaratan sesuai jenis ujikom
      parameters:
      - description: perpindahan_jabatan | kenaikan_tingkat
        in: query
        name: jenis_ujikom
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListDokumenPersyaratanResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List dokumen persyaratan
      tags:
      - ujikom
...
swagger: "2.0"